Fix inpainting, Use lru_cache, Fix Grid
