+++ preload.py
@@ -1,0 +1,7 @@
+def preload(parser):
+    parser.add_argument(
+        "--roop_loglevel",
+        default="INFO",
+        choices=["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"],
+        help="Set the log level (DEBUG, INFO, WARNING, ERROR, CRITICAL)",
+    )
