+++ scripts/faceswap.py
@@ -8,6 +8,7 @@ from scripts.roop_utils import models_utils
 from scripts.roop_postprocessing import upscaling
 from pprint import pprint
 import numpy as np
+import logging
 
 #Reload all the modules when using "apply and restart"
 #This is mainly done for development purposes
@@ -41,137 +42,15 @@ from PIL import Image
 
 from scripts.roop_utils.imgutils import (pil_to_cv2,convert_to_sd)
 
-from scripts.roop_logging import logger
+from scripts.roop_logging import logger, save_img_debug
 from scripts.roop_globals import VERSION_FLAG
 from scripts.roop_postprocessing.postprocessing_options import PostProcessingOptions
 from scripts.roop_postprocessing.postprocessing import enhance_image
+from scripts.faceswap_unit_settings import FaceSwapUnitSettings
+
 
 EXTENSION_PATH=os.path.join("extensions","sd-webui-roop")
 
-@dataclass
-class FaceSwapUnitSettings:
-    # The image given in reference
-    source_img: Union[Image.Image, str]
-    # The checkpoint file
-    source_face : str
-    # The batch source images
-    _batch_files: gr.components.File
-    # Will blend faces if True
-    blend_faces: bool
-    # Enable this unit
-    enable: bool
-    # Use same gender filtering
-    same_gender: bool
-
-    # If True, discard images with low similarity
-    check_similarity : bool 
-    # if True will compute similarity and add it to the image info
-    _compute_similarity :bool
-
-    # Minimum similarity against the used face (reference, batch or checkpoint)
-    min_sim: float
-    # Minimum similarity against the reference (reference or checkpoint if checkpoint is given)
-    min_ref_sim: float
-    # The face index to use for swapping
-    _faces_index: int
-    # Swap in the source image in img2img (before processing)
-    swap_in_source: bool
-    # Swap in the generated image in img2img (always on for txt2img)
-    swap_in_generated: bool
-
-    @staticmethod
-    def get_unit_configuration(unit: int, components):
-        fields_count = len(fields(FaceSwapUnitSettings))
-        return FaceSwapUnitSettings(
-            *components[unit * fields_count : unit * fields_count + fields_count]
-        )
-
-    @property
-    def faces_index(self):
-        """
-        Convert _faces_index from str to int 
-        """
-        faces_index = {
-            int(x) for x in self._faces_index.strip(",").split(",") if x.isnumeric()
-        }
-        if len(faces_index) == 0:
-            return {0}
-
-        return faces_index
-
-    @property
-    def compute_similarity(self) :
-        return self._compute_similarity or self.check_similarity
-
-    @property
-    def batch_files(self):
-        """
-        Return empty array instead of None for batch files
-        """
-        return self._batch_files or []
-    
-    @property
-    def reference_face(self) :
-        """
-        Extract reference face (only once and store it for the rest of processing).
-        Reference face is the checkpoint or the source image or the first image in the batch in that order.
-        """
-        if not hasattr(self,"_reference_face") :
-            if self.source_face and self.source_face != "None" :
-                with open(self.source_face, "rb") as file:
-                    try :
-                        logger.info(f"loading pickle {file.name}")
-                        face = Face(pickle.load(file))
-                        self._reference_face = face
-                    except Exception as e :
-                        logger.error("Failed to load checkpoint  : %s", e)
-            elif self.source_img is not None :
-                if isinstance(self.source_img, str):  # source_img is a base64 string
-                    if 'base64,' in self.source_img:  # check if the base64 string has a data URL scheme
-                        base64_data = self.source_img.split('base64,')[-1]
-                        img_bytes = base64.b64decode(base64_data)
-                    else:
-                        # if no data URL scheme, just decode
-                        img_bytes = base64.b64decode(self.source_img)
-                    self.source_img = Image.open(io.BytesIO(img_bytes))
-                source_img = pil_to_cv2(self.source_img)
-                self._reference_face =  swapper.get_or_default(swapper.get_faces(source_img), 0, None)  
-                if self._reference_face is None :
-                    logger.error("Face not found in reference image")  
-            else :
-                self._reference_face = None
-
-        if self._reference_face is None :
-            logger.error("You need at least one reference face")
-
-        return self._reference_face
-    
-    @property
-    def faces(self) :
-        """_summary_
-        Extract all faces (including reference face) to provide an array of faces
-        Only processed once.
-        """
-        if self.batch_files is not None and not hasattr(self,"_faces") :
-            self._faces = [self.reference_face] if self.reference_face is not None else []
-            for file in self.batch_files :
-                img = Image.open(file.name)
-                face = swapper.get_or_default(swapper.get_faces(pil_to_cv2(img)), 0, None)
-                if face is not None :
-                    self._faces.append(face)
-        return self._faces
-
-    @property
-    def blended_faces(self):
-        """
-        Blend the faces using the mean of all embeddings
-        """
-        if not hasattr(self,"_blended_faces") :
-            self._blended_faces = swapper.blend_faces(self.faces)
-            assert(all([not np.array_equal(self._blended_faces.embedding, face.embedding) for face in self.faces]) if len(self.faces) > 1 else True), "Blended faces cannot be the same as one of the face if len(face)>0"
-            assert(not np.array_equal(self._blended_faces.embedding,self.reference_face.embedding)  if len(self.faces) > 1 else True), "Blended faces cannot be the same as reference face if len(face)>0"
-
-        return self._blended_faces
 
 # Register the tab, done here to prevent it from being added twice
 script_callbacks.on_ui_tabs(faceswap_tab.on_ui_tabs)
@@ -284,12 +163,7 @@ class FaceSwapScript(scripts.Script):
                         logger.info(f"unit {i+1}> processed init images: {len(init_images)}, {len(result_infos)}")
                         unit.blend_faces = blend_config #restore blend config
 
-                p.init_images = init_images
-
-                # Apply mask :
-                for i,img in enumerate(p.init_images) :
-                    p.init_images[i] = imgutils.apply_mask(img, p, i)
-                        
+                p.init_images = init_images                        
 
     def process_image_unit(self, unit : FaceSwapUnitSettings, image: Image.Image, info = None, upscaled_swapper = False) -> List:
         """Process one image and return a List of (image, info) (one if blended, many if not).
@@ -324,6 +198,7 @@ class FaceSwapScript(scripts.Script):
                     logger.info("Use source face as reference face")
                     reference_face = src_face
 
+                save_img_debug(image, "Before swap")
                 result: swapper.ImageResult = swapper.swap_face(
                     reference_face,
                     src_face,
@@ -334,6 +209,8 @@ class FaceSwapScript(scripts.Script):
                     upscaled_swapper=upscaled_swapper,
                     compute_similarity=unit.compute_similarity
                 )
+                save_img_debug(result.image, "After swap")
+
                 if result.image is None :
                     logger.error("Result image is None")
                 if (not unit.check_similarity) or result.similarity and all([result.similarity.values()!=0]+[x >= unit.min_sim for x in result.similarity.values()]) and all([result.ref_similarity.values()!=0]+[x >= unit.min_ref_sim for x in result.ref_similarity.values()]):
@@ -371,11 +248,6 @@ class FaceSwapScript(scripts.Script):
             images = []
             infotexts = []
 
-            if self.keep_original_images:
-                # If we want to keep original images, we add all existing (including grid this time)
-                images = processed.images
-                infotexts = processed.infotexts
-
             for i,(img,info) in enumerate(zip(orig_images, orig_infotexts)): 
                 if any([u.enable for u in self.units]):
                     for unit_i, unit in enumerate(self.units):
@@ -384,12 +256,18 @@ class FaceSwapScript(scripts.Script):
                         batch_index = i%p.batch_size
                         if unit.enable :
                             if unit.swap_in_generated :
-                                swapped_images = self.process_image_unit(image=img, unit=unit, info=info, upscaled_swapper=self.upscaled_swapper_in_generated)
+                                img_to_swap = img
+                                
+                                swapped_images = self.process_image_unit(image=img_to_swap, unit=unit, info=info, upscaled_swapper=self.upscaled_swapper_in_generated)
                                 logger.info(f"{len(swapped_images)} images swapped")
                                 for swp_img, new_info in swapped_images :
                                     logger.info(f"unit {unit_i+1}> processed")
                                     if swp_img is not None :
+
+                                        save_img_debug(swp_img,"Before apply mask")
                                         swp_img = imgutils.apply_mask(swp_img, p, batch_index)
+                                        save_img_debug(swp_img,"After apply mask")
+
                                         try :   
                                             if self.postprocess_options is not None:
                                                 swp_img = enhance_image(swp_img, self.postprocess_options)
@@ -409,16 +287,21 @@ class FaceSwapScript(scripts.Script):
                                 images.append(img)
                                 infotexts.append(new_info)
 
+
             # Generate grid :
             if opts.return_grid and len(images) > 1:
                 # FIXME :Use sd method, not that if blended is not active, the result will be a bit messy.
-                grid = image_grid(images, p.batch_size)
+                grid = imgutils.create_square_image(images)
                 text = processed.infotexts[0]
                 infotexts.insert(0, text)
                 if opts.enable_pnginfo:
                     grid.info["parameters"] = text
                 images.insert(0, grid)
 
+            if self.keep_original_images:
+                # If we want to keep original images, we add all existing (including grid this time)
+                images += processed.images
+                infotexts += processed.infotexts
 
             processed.images = images
             processed.infotexts = infotexts
